CUS/pyproject.toml
[project]
name = "cus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomqtt>=2.5.1",
    "fastapi>=0.135.1",
    "pyserial-asyncio>=0.6",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
